Pong
Here is the start of a new coding game, pong! More like racquetball, since it is a single player game. There is currently just a rectangle platform that is able to move horizontally with the help of the mouse. There is also a ball but is stagnant. Hopfully in the next update, I will have it bouncing!